✨Refactor ElectionDescription to Manifest
- Move the files to a manifest file
- Keep election as just context and constants
